Implement  The Core Features #P2 (User Interest Domain)

+++ src/SAS.EventsService.Domain/Events/ValueObjects/EventInfo.cs
@@ -3,7 +3,7 @@
     public record EventInfo(
         string Title,
         string Summary,
-        int SentimentScore,
+        double SentimentScore,
         string SentimentLabel
         );
 }
